[DDH-13] Add Radio and Switch
