+++ src/app/page.js
@@ -1,4 +1,4 @@
-import MapFilter from "@/components/MapFilter";
+import MapFilter from "@/components/MapFilter/MapFilter";
 
 export default function Home() {
   return <div className="w-full h-full">
